+++ lib/providers/date_formats.dart
@@ -3,9 +3,9 @@ import 'package:intl/date_symbol_data_local.dart';
 import 'package:intl/intl.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import 'models/date_format_info.dart';
+import '../models/date_format_info.dart';
 
-part 'providers.g.dart';
+part 'date_formats.g.dart';
 
 @Riverpod(keepAlive: true)
 class SelectedDate extends _$SelectedDate {
