+++ lib/providers/locales.dart
