+++ lib/ui/locale_picker.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../locales.dart';
+import '../providers/locales.dart';
 
 class LocalePicker extends HookConsumerWidget {
   const LocalePicker({super.key});
